fixed game bugs and added nightmare button
Added a nightmare button that changes the level text and the button text and color. Nightmare mode is not actually created yet. The game has some bugs fixed not. You cannot change gamemodes or reset the game while the game is showing you the tile glows. You can only change the gamemode and reset once it's your turn