game over page
Added a game over page that has an animation and shows your highest level and what gamemode you did. So far there's only one design which is for the normal one (I haven't added a special version for nightmare). I also need to change the html file to gameOver instead of gameOverNormal and same for the script if I only am using one game over page and script.