added sounds and finished game over pages
Added two versions of the game over page, one for normal, and one for nightmare. I added sounds for when the player loses hearts. Normal has three hearts, nightmare has one. The hearts change color once it's lost. When you switch to nightmare there is a jumpscare sound, and when you lose, there's a jumpscare for nightmare. Both the normal and nightmare game over have the sound game over sound as the animation for the text plays. Nightmare's design looks more scary, normal's design is not. I also renamed the HTML, CSS, and JavaScript file for game over related to "gameOver" ending with it's type of file.