gameplay sounds
added sounds for when the tiles glow for nightmare and normal mode. It's different sounds for the different mode, and if you click the right tile, it plays the same sound again, otherwise a sound isn't played, but the heart loss sound plays. I added two sounds for nightmare, one is a deeper version of the drum, while the one I used is the regular version.