adjusted player click speed
Players can now tap the tiles faster. When the player taps a tile, the tile disappears faster so you can click the next one faster. Before, if you clicked on another tile while the previous one was still glowing, your next click would count as incorrect even though it was right, so I shortened the time on how long the tile glows when you tap it so you won't get the bug.