improved design
Added neon glow for the texts in normal and nightmare mode. The tiles with when no tiles has a glow has a nice box shadow, and the tiles glow differently depending on if you're on normal mode or nightmare mode. Both normal mode and nightmare mode have different error glow effect (when you hit the wrong tile).