Updated Reset Button
When you click on reset, it goes back to level 1 and you regain all your hearts. The game automatically starts. Reset did that before but the hearts did not reset, so that's what I fixed. Nightmare mode is the same, no change made (game goes back to level 1 and the game starts automatically. There's only one heart so there's no need for hearts to be reset, if you lost a heart in nightmare mode, the game ends).